[util] Generalize local stacked context
This is also useful for LLVM lowering, so let's move it to utils
